requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic[email]==2.5.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pillow==10.1.0
qrcode[pil]==7.4.2
apscheduler==3.10.4
